refactor: use ZipFile.open() instead of overwriting ZipFile
