+++ package_python_function/packager.py
@@ -1,11 +1,18 @@
+from __future__ import annotations
+
 import logging
+import os
 import shutil
+import time
 import zipfile
 from pathlib import Path
 from tempfile import NamedTemporaryFile
+from typing import TYPE_CHECKING
 
 from .python_project import PythonProject
-from .reproducible_zip import ReproducibleZipFile
+
+if TYPE_CHECKING:
+    from typing import Tuple
 
 logger = logging.getLogger(__name__)
 
@@ -39,14 +46,32 @@ class Packager:
     def zip_all_dependencies(self, target_path: Path) -> None:
         logger.info(f"Zipping to {target_path}...")
 
-        with ReproducibleZipFile(target_path, 'w', zipfile.ZIP_DEFLATED) as zip_file:
+        def date_time() -> Tuple[int, int, int, int, int, int]:
+            """Returns date_time value used to force overwrite on all ZipInfo objects. Defaults to
+            1980-01-01 00:00:00. You can set this with the environment variable SOURCE_DATE_EPOCH as an
+            integer value representing seconds since Epoch.
+            """
+            source_date_epoch = os.environ.get("SOURCE_DATE_EPOCH", None)
+            if source_date_epoch is not None:
+                return time.gmtime(int(source_date_epoch))[:6]
+            return (1980, 1, 1, 0, 0, 0)
+
+        with zipfile.ZipFile(target_path, "w", zipfile.ZIP_DEFLATED) as zip_file:
+
             def zip_dir(path: Path) -> None:
                 for item in path.iterdir():
                     if item.is_dir():
                         zip_dir(item)
                     else:
+                        zinfo = zipfile.ZipInfo.from_file(item)
+                        zinfo.date_time = date_time()
+                        zinfo.external_attr = 0o644 << 16
                         self._uncompressed_bytes += item.stat().st_size
-                        zip_file.write(item, item.relative_to(self.input_path))
+                        with (
+                            open(item, "rb") as src,
+                            zip_file.open(zinfo, "w") as dest,
+                        ):
+                            shutil.copyfileobj(src, dest, 1024 * 8)
 
             zip_dir(self.input_path)
